Initial implementation of stats reporting for GFS2 SRs
This is NOT the final implementation. This is a proof of concept.
Currently, an SR cannot be the owner of a datasource, nor it has
its own performance tab on XenCenter. Therefore, the GFS2 SR stats
get reported under the host's performance tab.

On attach, all hosts call start_stats(). However, the stats get
reported by the host that grabs the lock. If that process dies,
the host that grabs the lock starts reporting them.

Signed-off-by: Kostas Ladopoulos <Konstantinos.Ladopoulos@citrix.com>